requirements.txt
accelerate==0.19.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.5
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
cmake==3.26.3
comm==0.1.3
datasets==2.12.0
debugpy==1.6.7
decorator==4.4.2
diffusers==0.16.1
dill==0.3.6
executing==1.2.0
ffmpeg==1.4
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.14.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
lit==16.0.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.6
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
priority==2.0.0
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydub==0.25.1
pygame==2.4.0
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
quart==0.18.4
quart-cors==0.6.0
regex==2023.5.5
requests==2.30.0
responses==0.18.0
sentencepiece==0.1.99
six==1.16.0
soundfile==0.12.1
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
wsproto==1.2.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
